# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exam Result System API",
        "description": "API documentation for the high-performance Exam Result System.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/examinations/{examination}/results/compute": {
            "post": {
                "tags": [
                    "Results"
                ],
                "summary": "Compute all grades for an examination",
                "description": "Dispatches a fan-out queue architecture to compute grades for every student asynchronously.",
                "operationId": "9619d967fded5ab20b166ebfd1350333",
                "parameters": [
                    {
                        "name": "examination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Computation jobs dispatched to the queue"
                    }
                }
            }
        },
        "/api/examinations/{examination}/results/publish": {
            "post": {
                "tags": [
                    "Results"
                ],
                "summary": "Publish computed results for an examination",
                "description": "Moves all computed results to 'published' status and makes them visible to students. Requires X-Idempotency-Key header.",
                "operationId": "38496c2c938bfd10b9ffd04ff816c5c5",
                "parameters": [
                    {
                        "name": "examination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Results published with count"
                    }
                }
            }
        },
        "/api/enrollments/{enrollment}/result": {
            "get": {
                "tags": [
                    "Results"
                ],
                "summary": "View a single published result for an enrollment",
                "operationId": "542436886d0772475c50b30ebb6059c2",
                "parameters": [
                    {
                        "name": "enrollment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Published result"
                    },
                    "404": {
                        "description": "Result not available or not yet published"
                    }
                }
            }
        },
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "List all students",
                "operationId": "5a54d6df45ca455ad022d2f7b9da9731",
                "parameters": [
                    {
                        "name": "roll_number",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of students"
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "operationId": "aa0044bcc7e0cbaa736f0c20a4a6765c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "programme_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "roll_number": {
                                        "type": "string",
                                        "example": "CSE2026001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "batch_year": {
                                        "type": "integer",
                                        "example": 2026
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created"
                    }
                }
            }
        },
        "/api/students/{student}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Show a student with their enrolments",
                "operationId": "d67efa075fe7ae07b51c13d66feb5031",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student details with enrolments"
                    }
                }
            }
        }
    },
...
}